chore: move sign up hook out of user services
