+++ packages/apps/human-app/frontend/src/modules/signup/worker/hooks/use-sign-up-worker.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
 import omit from 'lodash/omit';
 import { browserAuthProvider } from '@/shared/contexts/browser-auth-provider';
-import type { SignUpDto } from '@/modules/worker/services/sign-up';
-import { useSignUpMutation } from '@/modules/worker/services/sign-up';
+import { type SignUpDto } from '../schema';
+import { useSignUpMutation } from './use-sign-up-mutation';
 
 export function useSignUpWorker() {
   const {
